changed java file to accept float as default instead of int, for now this is very scripd specific, it would serve us to generalize the inputs that can be taken in the future

+++ lib/mlkit.dart
@@ -257,7 +257,7 @@ class FirebaseModelInterpreter {
       {String remoteModelName,
       String localModelName,
       FirebaseModelInputOutputOptions inputOutputOptions,
-      Int32List inputBytes}) async {
+      Float64List inputBytes}) async {
     assert(remoteModelName != null || localModelName != null);
     try {
       dynamic results =
